// restore/lib/types.ts
export interface PaymentMethod {
  id: string;
  name: string;
  type: 'Dinheiro' | 'Pix' | 'Crédito' | 'Débito' | 'Fiado' | 'Voucher' | 'Outro';
  taxPercentage: number;
  taxFixed: number;
  active: boolean;
}